Add more general tree_extend/3 so all operations can use the same function.
