+++ src/lib.rs
@@ -1,3 +1,4 @@
+#[macro_use]
 pub mod app;
 
 pub fn get_tool_name() -> &'static str {
